Moving integration-tests to the parent module and adjust plugin dependencies
