+++ integration-tests/integration-tests-common/src/main/resources/org/acme/travels/timersCycle.bpmn
