+++ integration-tests/integration-tests-quarkus/src/main/resources/application.properties
