+++ integration-tests/integration-tests-quarkus/src/test/resources/application.properties
