+++ integration-tests/integration-tests-springboot/src/main/java/org/kie/kogito/KogitoApplication.java
