--- integration-tests/jobs-service-it/jobs-service-it-quarkus/src/test/java/org/kie/kogito/JobServiceQuarkusTestResource.java
@@ -1,50 +1,0 @@
-/**
- *  Copyright 2020 Red Hat, Inc. and/or its affiliates.
- *
- *  Licensed under the Apache License, Version 2.0 (the "License");
- *  you may not use this file except in compliance with the License.
- *  You may obtain a copy of the License at
- *
- *        http://www.apache.org/licenses/LICENSE-2.0
- *
- *  Unless required by applicable law or agreed to in writing, software
- *  distributed under the License is distributed on an "AS IS" BASIS,
- *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *  See the License for the specific language governing permissions and
- *  limitations under the License.
- */
-package org.kie.kogito;
-
-import org.kie.kogito.resources.ConditionalQuarkusTestResource;
-import org.kie.kogito.testcontainers.JobServiceContainer;
-
-/**
- * Infinispan quarkus resource that works within the test lifecycle.
- *
- */
-public class JobServiceQuarkusTestResource extends ConditionalQuarkusTestResource {
-
-    public static final String CONFIG_PROPERTY_TO_EXPOSE = "kogito.jobs-service.url";
-
-    public JobServiceQuarkusTestResource() {
-        super(new JobServiceContainer());
-    }
-
-    @Override
-    protected String getKogitoProperty() {
-        return CONFIG_PROPERTY_TO_EXPOSE;
-    }
-
-    @Override
-    protected String getKogitoPropertyValue() {
-        return "http://localhost:" + getTestResource().getMappedPort();
-    }
-
-    public static class Conditional extends JobServiceQuarkusTestResource {
-
-        public Conditional() {
-            super();
-            enableConditional();
-        }
-    }
-}
