--- integration-tests/jobs-service-it/jobs-service-it-springboot/src/test/java/org/kie/kogito/ProcessTimerIT.java
@@ -1,34 +1,0 @@
-/**
- *  Copyright 2020 Red Hat, Inc. and/or its affiliates.
- *
- *  Licensed under the Apache License, Version 2.0 (the "License");
- *  you may not use this file except in compliance with the License.
- *  You may obtain a copy of the License at
- *
- *        http://www.apache.org/licenses/LICENSE-2.0
- *
- *  Unless required by applicable law or agreed to in writing, software
- *  distributed under the License is distributed on an "AS IS" BASIS,
- *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *  See the License for the specific language governing permissions and
- *  limitations under the License.
- */
-package org.kie.kogito;
-
-import org.junit.jupiter.api.BeforeAll;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.annotation.DirtiesContext;
-import org.springframework.test.context.ContextConfiguration;
-import org.springframework.util.SocketUtils;
-
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT, classes = {KogitoApplication.class})
-@ContextConfiguration(initializers = JobServiceSpringBootTestResource.class)
-@DirtiesContext(classMode = DirtiesContext.ClassMode.AFTER_CLASS)
-public class ProcessTimerIT extends BaseProcessTimerIT {
-
-    @BeforeAll
-    public static void beforeAll() {
-        BaseProcessTimerIT.beforeAll(() -> SocketUtils.findAvailableTcpPort());
-        System.setProperty("server.port", String.valueOf(BaseProcessTimerIT.httpPort));
-    }
-}
