+++ integration-tests/pom.xml
@@ -11,6 +11,70 @@
   <packaging>pom</packaging>
 
   <modules>
-    <module>jobs-service-it</module>
+    <module>integration-tests-common</module>
+    <module>integration-tests-quarkus</module>
+    <module>integration-tests-springboot</module>
   </modules>
+
+  <dependencyManagement>
+    <dependencies>
+      <dependency>
+        <groupId>org.kie.kogito</groupId>
+        <artifactId>kogito-bom</artifactId>
+        <version>${project.version}</version>
+        <type>pom</type>
+        <scope>import</scope>
+      </dependency>
+      <dependency>
+        <groupId>io.quarkus</groupId>
+        <artifactId>quarkus-bom</artifactId>
+        <version>${version.io.quarkus}</version>
+        <type>pom</type>
+        <scope>import</scope>
+      </dependency>
+      <dependency>
+        <groupId>org.kie.kogito</groupId>
+        <artifactId>integration-tests-common</artifactId>
+        <version>${project.version}</version>
+      </dependency>
+    </dependencies>
+  </dependencyManagement>
+
+  <build>
+    <pluginManagement>
+      <plugins>
+        <plugin>
+          <groupId>org.apache.maven.plugins</groupId>
+          <artifactId>maven-failsafe-plugin</artifactId>
+          <dependencies>
+            <!-- ensure jobs-service module was built, but not added in the classpath -->
+            <dependency>
+              <groupId>org.kie.kogito</groupId>
+              <artifactId>jobs-service</artifactId>
+              <version>${project.version}</version>
+              <scope>runtime</scope>
+            </dependency>
+          </dependencies>
+          <configuration>
+            <systemPropertyVariables>
+              <container.image.infinispan>${container.image.infinispan}</container.image.infinispan>
+              <container.image.jobs-service>${container.image.jobs-service}</container.image.jobs-service>
+              <container.image.keycloak>${container.image.keycloak}</container.image.keycloak>
+            </systemPropertyVariables>
+            <classesDirectory>${project.build.outputDirectory}</classesDirectory>
+          </configuration>
+          <executions>
+            <execution>
+              <id>default</id>
+              <phase>test</phase>
+              <goals>
+                <goal>integration-test</goal>
+                <goal>verify</goal>
+              </goals>
+            </execution>
+          </executions>
+        </plugin>
+      </plugins>
+    </pluginManagement>
+  </build>
 </project>
